django_core/requirements/base.txt
argon2-cffi==21.3.0
celery==5.2.7
django==4.1
django-filter==22.1
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-constance==2.9.1
django-colorfield==0.7.2
django-picklefield==3.1
django-import-export==3.0.1
django-filebrowser-no-grappelli==4.0.1
django-environ==0.9.0
django-jazzmin==2.5.0
django-lifecycle==1.0.0
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
django-tinymce==3.5.0
djoser==2.1.0
gunicorn==20.1.0
transliterate==1.10.2
whitenoise==6.2.0
psycopg2-binary==2.9.5
drf_spectacular==0.23.1
